+++ src/FestivalPOS/src/checkout/cash-payment.ts
@@ -45,11 +45,13 @@ export class CashPayment {
 
     async cancel() {
         try {
-            this.progress.busy("Sletter ordre", faTrash);
+            if (!this.progress.tryBusy("Sletter ordre", faTrash)) {
+                return;
+            }
 
             await this.api.deleteOrder(this.order.id).send();
 
-            this.progress.done();
+            await this.progress.done();
 
             this.router.navigate("/sale");
         }
@@ -64,14 +66,16 @@ export class CashPayment {
         }
 
         try {
-            this.progress.busy("Registrerer betaling", faCashRegister);
+            if (!this.progress.tryBusy("Registrerer betaling", faCashRegister)) {
+                return;
+            }
 
             const payment = await this.api.createPayment(this.order.id, {
                 method: "cash",
                 amount: this.amountDue
             }).transfer();
 
-            this.progress.done();
+            await this.progress.done();
 
             this.router.navigateToRoute("receipt", {
                 orderId: payment.orderId,
